broke up trees into classes
the trees now consist of the abstract tree class (not implemented), the tree base class which has an array of trunks, and the tree trunk class which is instantiated with each growth and tracks the base class. each tree instance tracks when it is growing and clicks don't cause growth. the tree base as a temporary 'silly' function that causes the tree to shift around